Move the `hooks-common` module under the `@bayou` namespace.
