+++ local-modules/doc-common/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@bayou/codec": "local",
     "@bayou/deps-quill-common": "local",
-    "hooks-common": "local",
+    "@bayou/hooks-common": "local",
     "ot-common": "local",
     "@bayou/typecheck": "local",
     "@bayou/util-common": "local"
